+++ changelog.md
@@ -5,6 +5,7 @@
 ## v0.1.0-beta.4
 
 - Fixed crashes when garbage-collecting node/object userdata.
+- Fixed tiny ImGui on macOS Retina with improved scaling and fonts (forked matcool/gd-imgui-cocos).
 
 ## v0.1.0-beta.3